Hometask 5

+++ Ivanova Natasha/homework 5/sqhema_cafe_db.sql
@@ -16,14 +16,14 @@ first_name VARCHAR(30) NOT NULL,
 second_name VARCHAR(30) NOT NULL,
 vk_id VARCHAR(30) NOT NULL,
 registration_dttm TIMESTAMP NOT NULL,
-status_id BIGINT(20) UNSIGNED NOT NULL,
+status_id INT UNSIGNED NOT NULL,
 
 FOREIGN KEY (status_id) REFERENCES statuses(status_id) ON DELETE CASCADE
 );
 
 CREATE TABLE orders (
 order_id SERIAL PRIMARY KEY,
-client_id BIGINT(20) UNSIGNED NOT NULL,
+client_id INT UNSIGNED NOT NULL,
 total_price INT UNSIGNED NOT NULL DEFAULT 0,
 
 FOREIGN KEY (client_id) REFERENCES clients(client_id) ON DELETE CASCADE
@@ -47,8 +47,8 @@ first_name VARCHAR(30) NOT NULL,
 second_name VARCHAR(30) NOT NULL,
 vk_id VARCHAR(30) NOT NULL,
 first_day_dttm TIMESTAMP NOT NULL,
-shift_id BIGINT(20) UNSIGNED NOT NULL,
-role_id BIGINT(20) UNSIGNED NOT NULL,
+shift_id INT UNSIGNED NOT NULL,
+role_id INT UNSIGNED NOT NULL,
 
 FOREIGN KEY (shift_id) REFERENCES shifts(shift_id) ON DELETE CASCADE
 FOREIGN KEY (role_id) REFERENCES roles(role_id) ON DELETE CASCADE
@@ -65,15 +65,15 @@ CREATE TABLE products (
 product_id SERIAL PRIMARY KEY,
 name VARCHAR(100) NOT NULL,
 price INT UNSIGNED  NOT NULL DEFAULT 0,
-provider_id BIGINT(20) UNSIGNED NOT NULL,
+provider_id INT UNSIGNED NOT NULL,
 
 FOREIGN KEY (provider_id) REFERENCES providers(provider_id) ON DELETE CASCADE
 );
 
 CREATE TABLE order_item_link (
 order_item_link_id SERIAL PRIMARY KEY,
-item_id BIGINT(20) UNSIGNED NOT NULL,
-order_id BIGINT(20) UNSIGNED NOT NULL,
+item_id INT UNSIGNED NOT NULL,
+order_id INT UNSIGNED NOT NULL,
 quantity SMALLINT UNSIGNED NOT NULL,
 
 FOREIGN KEY (order_id) REFERENCES orders(order_id) ON DELETE CASCADE,
@@ -82,8 +82,8 @@ FOREIGN KEY (item_id) REFERENCES items(item_id) ON DELETE CASCADE
 
 CREATE TABLE product_item_link (
 product_item_link_id SERIAL PRIMARY KEY,
-product_id BIGINT(20) UNSIGNED NOT NULL,
-item_id BIGINT(20) UNSIGNED NOT NULL,
+product_id INT UNSIGNED NOT NULL,
+item_id INT UNSIGNED NOT NULL,
 quantity SMALLINT UNSIGNED NOT NULL,
 
 FOREIGN KEY (product_id) REFERENCES products(product_id) ON DELETE CASCADE,
